+++ coeus-code/src/main/java/org/kuali/coeus/common/budget/framework/distribution/BudgetCostShare.java
@@ -13,14 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.budget.distributionincome;
+package org.kuali.coeus.common.budget.framework.distribution;
 
+import org.kuali.coeus.common.budget.framework.core.BudgetDistributionAndIncomeComponent;
 import org.kuali.coeus.sys.api.model.ScaleTwoDecimal;
 import org.kuali.coeus.propdev.impl.hierarchy.HierarchyMaintainable;
 
-/**
- * 
- */
 public class BudgetCostShare extends BudgetDistributionAndIncomeComponent implements HierarchyMaintainable {
     private static final long serialVersionUID = 6199797319981907016L;
