+++ coeus-code/src/main/java/org/kuali/coeus/common/budget/framework/distribution/BudgetDistributionService.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.budget.distributionincome;
+package org.kuali.coeus.common.budget.framework.distribution;
 
 import org.kuali.kra.budget.core.Budget;
 
-public interface BudgetDistributionAndIncomeService {
+public interface BudgetDistributionService {
 
     /**
      * This method initializes both the Cost Sharing and Unrecovered F&A defaults
