+++ coeus-code/src/main/java/org/kuali/coeus/common/budget/framework/income/BudgetProjectIncomeRuleImpl.java
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.budget.distributionincome;
+package org.kuali.coeus.common.budget.framework.income;
 
+
+import org.kuali.coeus.common.budget.impl.core.ValidationHelper;
 
 /**
  * ProcessDefinitionDefinitiones Budget Project Income rules
