+++ coeus-code/src/main/java/org/kuali/coeus/common/budget/impl/distribution/BudgetCostShareAllocationEvent.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.budget.distributionincome;
+package org.kuali.coeus.common.budget.impl.distribution;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.kuali.coeus.common.budget.framework.distribution.BudgetCostShareAllocationRule;
 import org.kuali.coeus.sys.framework.rule.KcDocumentEventBase;
 import org.kuali.kra.budget.document.BudgetDocument;
 import org.kuali.rice.krad.rules.rule.BusinessRule;
