+++ coeus-code/src/main/java/org/kuali/kra/budget/core/Budget.java
@@ -35,10 +35,10 @@ import org.kuali.coeus.sys.framework.service.KcServiceLocator;
 import org.kuali.kra.bo.InstituteLaRate;
 import org.kuali.kra.bo.InstituteRate;
 import org.kuali.kra.budget.calculator.BudgetCalculationService;
-import org.kuali.kra.budget.distributionincome.BudgetCostShare;
-import org.kuali.kra.budget.distributionincome.BudgetDistributionAndIncomeComponent;
-import org.kuali.kra.budget.distributionincome.BudgetProjectIncome;
-import org.kuali.kra.budget.distributionincome.BudgetUnrecoveredFandA;
+import org.kuali.coeus.common.budget.framework.distribution.BudgetCostShare;
+import org.kuali.coeus.common.budget.framework.core.BudgetDistributionAndIncomeComponent;
+import org.kuali.coeus.common.budget.framework.income.BudgetProjectIncome;
+import org.kuali.coeus.common.budget.framework.distribution.BudgetUnrecoveredFandA;
 import org.kuali.kra.budget.document.BudgetDocument;
 import org.kuali.kra.budget.nonpersonnel.BudgetFormulatedCostDetail;
 import org.kuali.kra.budget.nonpersonnel.BudgetLineItem;
