+++ coeus-code/src/main/java/org/kuali/kra/budget/document/BudgetDocumentRule.java
@@ -16,8 +16,15 @@
 package org.kuali.kra.budget.document;
 
 import org.apache.commons.lang3.StringUtils;
+import org.kuali.coeus.common.budget.framework.distribution.*;
 import org.kuali.coeus.common.budget.framework.period.BudgetPeriod;
+import org.kuali.coeus.common.budget.framework.income.AddBudgetProjectIncomeEvent;
+import org.kuali.coeus.common.budget.framework.income.AddBudgetProjectIncomeRule;
+import org.kuali.coeus.common.budget.framework.income.BudgetProjectIncome;
+import org.kuali.coeus.common.budget.framework.income.BudgetProjectIncomeRuleImpl;
+import org.kuali.coeus.common.budget.impl.distribution.BudgetCostShareRuleImpl;
 import org.kuali.coeus.common.budget.impl.period.*;
+import org.kuali.coeus.common.budget.framework.distribution.BudgetUnrecoveredFandAAuditRule;
 import org.kuali.coeus.common.framework.custom.KcDocumentBaseAuditRule;
 import org.kuali.coeus.sys.api.model.ScaleTwoDecimal;
 import org.kuali.kra.award.budget.AwardBudgeCostTotalAuditRule;
@@ -25,7 +32,6 @@ import org.kuali.kra.award.budget.AwardBudgetBudgetTypeAuditRule;
 import org.kuali.kra.award.budget.AwardBudgetCostLimitAuditRule;
 import org.kuali.kra.award.budget.document.AwardBudgetDocument;
 import org.kuali.kra.budget.core.BudgetParent;
-import org.kuali.kra.budget.distributionincome.*;
 import org.kuali.kra.budget.nonpersonnel.BudgetExpensesAuditRule;
 import org.kuali.kra.budget.nonpersonnel.BudgetLineItem;
 import org.kuali.kra.budget.personnel.BudgetPersonnelAuditRule;
@@ -56,7 +62,7 @@ import java.util.List;
 public class BudgetDocumentRule extends CostShareRuleResearchDocumentBase implements AddBudgetPeriodRule, AddBudgetCostShareRule, AddBudgetProjectIncomeRule, SaveBudgetPeriodRule, DeleteBudgetPeriodRule, GenerateBudgetPeriodRule, DocumentAuditRule, SyncModularBudgetRule {
 
     /** 
-     * @see org.kuali.kra.budget.distributionincome.AddBudgetCostShareRule#processAddBudgetCostShareBusinessRules(org.kuali.kra.budget.distributionincome.AddBudgetCostShareEvent)
+     * @see org.kuali.coeus.common.budget.framework.distribution.AddBudgetCostShareRule#processAddBudgetCostShareBusinessRules(org.kuali.coeus.common.budget.framework.distribution.AddBudgetCostShareEvent)
      */
     public boolean processAddBudgetCostShareBusinessRules(AddBudgetCostShareEvent addBudgetCostShareEvent) {
         return new BudgetCostShareRuleImpl().processAddBudgetCostShareBusinessRules(addBudgetCostShareEvent);
