+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/BudgetCostShare.xml
@@ -2,7 +2,7 @@
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p" xsi:schemaLocation="http://www.springframework.org/schema/beans         http://www.springframework.org/schema/beans/spring-beans.xsd">
   <bean id="BudgetCostShare" parent="BudgetCostShare-parentBean" />
   <bean id="BudgetCostShare-parentBean" abstract="true" parent="BusinessObjectEntry">
-    <property name="businessObjectClass" value="org.kuali.kra.budget.distributionincome.BudgetCostShare" />
+    <property name="businessObjectClass" value="org.kuali.coeus.common.budget.framework.distribution.BudgetCostShare" />
 		<property name="objectLabel" value="Budget Cost Share" />
     <property name="titleAttribute" value="budgetCostShare" />
     <property name="helpDefinition" >
