+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/BudgetProjectIncome.xml
@@ -2,7 +2,7 @@
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p" xsi:schemaLocation="http://www.springframework.org/schema/beans         http://www.springframework.org/schema/beans/spring-beans.xsd">
   <bean id="BudgetProjectIncome" parent="BudgetProjectIncome-parentBean" />
   <bean id="BudgetProjectIncome-parentBean" abstract="true" parent="BusinessObjectEntry">
-    <property name="businessObjectClass" value="org.kuali.kra.budget.distributionincome.BudgetProjectIncome" />
+    <property name="businessObjectClass" value="org.kuali.coeus.common.budget.framework.income.BudgetProjectIncome" />
 		<property name="objectLabel" value="Budget Project Income" />
     <property name="titleAttribute" value="budgetPeriod" />
     <property name="helpDefinition" >
