+++ coeus-code/src/main/resources/org/kuali/kra/datadictionary/BudgetUnrecoveredFandA.xml
@@ -2,7 +2,7 @@
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:p="http://www.springframework.org/schema/p" xsi:schemaLocation="http://www.springframework.org/schema/beans         http://www.springframework.org/schema/beans/spring-beans.xsd">
   <bean id="BudgetUnrecoveredFandA" parent="BudgetUnrecoveredFandA-parentBean" />
   <bean id="BudgetUnrecoveredFandA-parentBean" abstract="true" parent="BusinessObjectEntry">
-    <property name="businessObjectClass" value="org.kuali.kra.budget.distributionincome.BudgetUnrecoveredFandA" />
+    <property name="businessObjectClass" value="org.kuali.coeus.common.budget.framework.distribution.BudgetUnrecoveredFandA" />
 		<property name="objectLabel" value="Budget Unrecovered Fand A" />
     <property name="titleAttribute" value="documentComponentId" />
     <property name="helpDefinition" >
