+++ coeus-code/src/test/java/org/kuali/kra/budget/bo/BudgetProjectIncomeTest.java
@@ -20,7 +20,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.kuali.kra.budget.core.Budget;
-import org.kuali.kra.budget.distributionincome.BudgetProjectIncome;
+import org.kuali.coeus.common.budget.framework.income.BudgetProjectIncome;
 import org.kuali.coeus.common.budget.framework.period.BudgetPeriod;
 import org.kuali.coeus.sys.framework.util.DateUtils;
 import org.kuali.coeus.sys.api.model.ScaleTwoDecimal;
