+++ coeus-code/src/test/java/org/kuali/kra/budget/rules/AddBudgetProjectIncomeRuleTest.java
@@ -19,15 +19,14 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.kuali.kra.budget.distributionincome.AddBudgetProjectIncomeEvent;
-import org.kuali.kra.budget.distributionincome.AddBudgetProjectIncomeRule;
-import org.kuali.kra.budget.distributionincome.BudgetProjectIncome;
-import org.kuali.kra.budget.distributionincome.BudgetProjectIncomeRuleImpl;
+import org.kuali.coeus.common.budget.framework.income.AddBudgetProjectIncomeEvent;
+import org.kuali.coeus.common.budget.framework.income.AddBudgetProjectIncomeRule;
+import org.kuali.coeus.common.budget.framework.income.BudgetProjectIncome;
+import org.kuali.coeus.common.budget.framework.income.BudgetProjectIncomeRuleImpl;
 import org.kuali.coeus.sys.api.model.ScaleTwoDecimal;
 import org.kuali.rice.krad.util.ErrorMessage;
 import org.kuali.rice.krad.util.GlobalVariables;
 import org.kuali.rice.krad.util.MessageMap;
-import org.springframework.util.AutoPopulatingList;
 
 import java.util.List;
 import java.util.Properties;
