+++ coeus-help/src/main/resources/AKLinks/TX__2044.js
@@ -1,1 +1,1 @@
-﻿d2hStoreMenuItems("TX__2044", [["Documents/costsharing.htm", "right", "Cost Sharing"],["Documents/costsharing1.htm", "right", "Cost Sharing"],["Documents/costsharing2.htm", "right", "Cost Sharing"],["Documents/costsharing3.htm", "right", "Cost Sharing"]]);
+﻿d2hStoreMenuItems("TX__2044", [["Documents/distribution.htm", "right", "Cost Sharing"],["Documents/costsharing1.htm", "right", "Cost Sharing"],["Documents/costsharing2.htm", "right", "Cost Sharing"],["Documents/costsharing3.htm", "right", "Cost Sharing"]]);
